+++ src/components/NotificationWidget/index.js
@@ -1,11 +1,17 @@
 import * as React from 'react'
 import { css, cx } from 'react-emotion'
 import { connect } from 'react-redux'
+import {
+  Modal,
+  PopupBody,
+  PopupFooter,
+  CopyToClipboard,
+  Button
+} from '@tarantool.io/ui-kit'
 import bell from './notification.svg'
 import { clearNotifications, hideNotificationList, showNotificationList } from '../../store/actions/notifications'
 import type { NotificationItem } from '../../store/reducers/notifications'
 import Notification from '../Notification'
-import Button from '../Button'
 import { isExistsHiddenNonRead, selectHiddenNotifications } from '../../store/selectors'
 import { AutoScroll } from '../AutoScroll'
 
@@ -87,6 +93,10 @@ type NotificationWidgetProps = {
 }
 
 class NotificationWidget extends React.PureComponent<NotificationWidgetProps> {
+  state = {
+    details: null,
+    showDetailsModal: false
+  }
   refEl = null
   clickHandler = e => {
     if (this.refEl !== e.target && !this.refEl.contains(e.target)) {
@@ -105,10 +115,42 @@ class NotificationWidget extends React.PureComponent<NotificationWidgetProps> {
     }
   }
 
+  handleDetailsClick = details => {
+    this.setState({ details, showDetailsModal: true });
+  }
+
+  renderDetailsModal = () => {
+    const {
+      showDetailsModal,
+      details
+    } = this.state;
+    return (
+      <Modal
+        title='Details'
+        visible={showDetailsModal}
+        onClose={() => this.setState({ showDetailsModal: false })}
+      >
+        <PopupBody>{details}</PopupBody>
+        <PopupFooter
+          controls={[
+            <CopyToClipboard
+              key="copy"
+              content={details}
+            >
+              Copy details
+            </CopyToClipboard>,
+            <Button key="close" text='Close' onClick={() => this.setState({ showDetailsModal: false })} />
+          ]}
+        />
+      </Modal>
+    );
+  }
+
   render(): React.ReactNode {
     const { dispatch, notifications, showList, active } = this.props
     return (
       <div className={styles.container} ref={r => (this.refEl = r)}>
+        {this.renderDetailsModal()}
         <div
           className={cx(styles.bell, active ? styles.bellActive : null)}
           onClick={() => {
@@ -124,7 +166,13 @@ class NotificationWidget extends React.PureComponent<NotificationWidgetProps> {
                 <div className={styles.notificationInner}>
                   {notifications.length === 0 ? <span className={styles.noNotification}>No notifications</span> : null}
                   {notifications.map(x => (
-                    <Notification key={x.uuid} className={styles.listItem} {...x} isShort={true} />
+                    <Notification
+                      key={x.uuid}
+                      className={styles.listItem}
+                      {...x}
+                      onDetailsClick={details => this.handleDetailsClick(details)}
+                      isShort={true}
+                    />
                   ))}
                 </div>
               </AutoScroll>