+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove base64 inline Roboto fonts & add Inter font
 - Update set title when change location
 - Passing data to frontend with `front.set_variable()`
+- Introduced `@tarantool.io/ui-kit`
+- Added `details` prop to notifications
 
 ## [7.0.0]